Add 3D matrix addition and tensor multiply
Add CPU and CUDA examples for 3D matrix addition and tensor multiplication. New files: 3D Matrix Addition/3d.cpp (CPU implementation) and Tensor Multiplication/mul.cu (CUDA matrix multiply); also add compiled binaries (3d, 3d2, mul). Update 3D Matrix Addition/3d.cu to use long/unsigned long for indices and sizes (and corresponding %li printf specifiers) to support larger dimensions and adjust allocation sizes and loop counters accordingly.

+++ 3D Matrix Addition/3d.cu
@@ -1,13 +1,13 @@
 #include <iostream>
 #include <cuda_runtime.h>
 
-__global__ void add(int* a, int* b, int* c, int n){
-    int row = blockDim.x * blockIdx.x + threadIdx.x;
-    int col = blockDim.y * blockIdx.y + threadIdx.y;
-    int dep = blockDim.z * blockIdx.z + threadIdx.z;
+__global__ void add(long* a, long* b, long* c, int n){
+    long row = blockDim.x * blockIdx.x + threadIdx.x;
+    long col = blockDim.y * blockIdx.y + threadIdx.y;
+    long dep = blockDim.z * blockIdx.z + threadIdx.z;
 
     if (row < n && col < n && dep < n){
-        int idx = row*n*n + col*n + dep;
+        long idx = row*n*n + col*n + dep;
         a[idx] = idx*2;
         b[idx] = idx*3;
 
@@ -16,7 +16,7 @@ __global__ void add(int* a, int* b, int* c, int n){
 }
 
 int main(void){
-    unsigned int n, k;
+    unsigned long n, k;
     std::cout << "Enter matrix dimension (n): ";
     std::cin >> n;
     if (n>10){
@@ -28,9 +28,9 @@ int main(void){
 
     dim3 numBlocks((n + threadCount.x - 1)/threadCount.x, (n + threadCount.y - 1)/threadCount.y, (n + threadCount.z - 1)/threadCount.z);
 
-    int* a, *b, *c;
+    long* a, *b, *c;
     
-    size_t size = n*n*n*sizeof(int);
+    size_t size = n*n*n*sizeof(long);
 
     cudaMallocManaged(&a, size);
     cudaMallocManaged(&b, size);
@@ -40,14 +40,14 @@ int main(void){
 
     cudaDeviceSynchronize();
 
-    printf("The first sum is: %i + %i = %i\n", a[0], b[0], c[0]);
-    printf("The second sum is: %i + %i = %i\n", a[1], b[1], c[1]);
-    printf("The third sum is: %i + %i = %i\n", a[2], b[2], c[2]);
-    printf("The fourth sum is: %i + %i = %i\n", a[3], b[3], c[3]);
-    printf("The fifth sum is: %i + %i = %i\n", a[4], b[4], c[4]);
+    printf("The first sum is: %li + %li = %li\n", a[0], b[0], c[0]);
+    printf("The second sum is: %li + %li = %li\n", a[1], b[1], c[1]);
+    printf("The third sum is: %li + %li = %li\n", a[2], b[2], c[2]);
+    printf("The fourth sum is: %li + %li = %li\n", a[3], b[3], c[3]);
+    printf("The fifth sum is: %li + %li = %li\n", a[4], b[4], c[4]);
     printf("The sums are: ");
-    for (int i = 0; i<n*n*n; i++){
-        printf("%i, ", c[i]);
+    for (long i = 0; i<n*n*n; i++){
+        printf("%li, ", c[i]);
     }
 
     cudaFree(a);
